+++ glue/core/data_factories/fits.py
@@ -17,7 +17,7 @@ def is_fits(filename):
     try:
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
-            with fits.open(filename, ignore_missing_end=True):
+            with fits.open(filename, ignore_missing_end=True, mode='denywrite'):
                 return True
     except IOError:
         return False
@@ -57,7 +57,7 @@ def fits_reader(source, auto_merge=False, exclude_exts=None, label=None):
         hdulist = source
         close_hdulist = False
     else:
-        hdulist = fits.open(source, ignore_missing_end=True)
+        hdulist = fits.open(source, ignore_missing_end=True, mode='denywrite')
         hdulist.verify('fix')
         close_hdulist = True
 
@@ -173,7 +173,7 @@ def is_casalike(filename, **kwargs):
 
     if not is_fits(filename):
         return False
-    with fits.open(filename, ignore_missing_end=True) as hdulist:
+    with fits.open(filename, ignore_missing_end=True, mode='denywrite') as hdulist:
         if len(hdulist) != 1:
             return False
         if hdulist[0].header['NAXIS'] != 4:
@@ -202,7 +202,7 @@ def casalike_cube(filename, **kwargs):
     if 'ignore_missing_end' not in kwargs:
         kwargs['ignore_missing_end'] = True
 
-    with fits.open(filename, **kwargs) as hdulist:
+    with fits.open(filename, mode='denywrite', **kwargs) as hdulist:
         array = hdulist[0].data
         header = hdulist[0].header
     result.coords = coordinates_from_header(header)